queue: slight code cleanup
